Update docker for USB storage

+++ wdpk/example/common.sh
@@ -6,15 +6,19 @@ APP_NAME=example
 LOG=/tmp/apkg_$APP_NAME.log
 APKG_PATH=$1
 
-# Delete existing log if it hasn't been modified in the last week
-find "$LOG" -mtime +7 -delete
+# If we're not connected to a terminal, set up logging to $LOG
+if ! [ -t 0 ] && [ -z "$APKG_LOGGING_SETUP" ]; then
+    # Delete existing log if it hasn't been modified in the last week
+    find "$LOG" -mtime +7 -delete
 
-# Ensure /dev/fd works (for process substitution)
-ln -sfn /proc/self/fd /dev/fd
+    # Ensure /dev/fd works (for process substitution)
+    ln -sfn /proc/self/fd /dev/fd
 
-# Tee STDOUT and STDERR to log file
-exec > >(tee -ia $LOG) 2>&1
+    # Tee STDOUT and STDERR to log file
+    exec > >(tee -ia $LOG) 2>&1
 
-echo "$(date -Iseconds) APKG_DEBUG: $0 $*"
+    echo "$(date -Iseconds) APKG_DEBUG: $0 $*"
+    export APKG_LOGGING_SETUP=true
 
-set -xo pipefail
+    set -xo pipefail
+fi
